fix discourse topic display
